+++ cmd/sense/main.go
@@ -14,7 +14,7 @@ func main() {
 	var (
 		cli       = cmdline.NewBasicParser()
 		root, _   = os.Getwd()
-		pause     = cli.Option("-i, --interval").Duration("2s")
+		interval  = cli.Option("-i, --interval").Duration("2s")
 		recursive = cli.Flag("-r, --recursive")
 		script    = cli.Option("-s, --script").String("./.onchange.sh")
 		writeEx   = cli.Flag("-w, --write-example-script")
@@ -35,7 +35,7 @@ func main() {
 		os.Stdout.Write(out)
 	})
 	s.Recursive = recursive
-	s.Pause = pause
+	s.Interval = interval
 	s.Run(context.Background())
 }